requirements.txt
# dependencies
colorama
openai
python-dotenv
uvicorn
pydantic
fastapi
python-multipart
langchain
langchain-openai
langchain_community
requests
SQLAlchemy
pytest
pytest-asyncio
langchain-community
langchain-experimental
langchainhub
langsmith
PyPDF2
langchain_qdrant
psycopg2
redis
starlette
celery
passlib[bcrypt]
pydantic[email]
langgraph
psycopg
psycopg-pool
langgraph-checkpoint-postgres
IPython
kafka
PyJWT
aiokafka
google-auth
google-api-python-client
python-dateutil
asyncpg